restructured cipher client to sub folders
