+++ crates/bitwarden-vault/src/cipher/cipher_client/mod.rs
@@ -10,13 +10,13 @@ use wasm_bindgen::prelude::*;
 
 use super::EncryptionContext;
 use crate::{
-    cipher::cipher::DecryptCipherListResult,
-    create::{create_cipher, CipherAddEditRequest, CreateCipherError},
-    edit::{edit_cipher, EditCipherError},
-    get_list::{get_cipher, list_ciphers, list_ciphers_with_failures, GetCipherError},
-    Cipher, CipherError, CipherListView, CipherView, DecryptError, EncryptError,
-    Fido2CredentialFullView,
+    cipher::cipher::DecryptCipherListResult, Cipher, CipherError, CipherListView, CipherView,
+    DecryptError, EncryptError, Fido2CredentialFullView,
 };
+
+pub mod create;
+pub mod edit;
+pub mod get;
 
 #[allow(missing_docs)]
 #[cfg_attr(feature = "wasm", wasm_bindgen)]
@@ -180,68 +180,6 @@ impl CiphersClient {
         let key_store = self.client.internal.get_key_store();
         let decrypted_key = cipher_view.decrypt_fido2_private_key(&mut key_store.context())?;
         Ok(decrypted_key)
-    }
-
-    /// Get all ciphers from state and decrypt them to a list of [CipherView].
-    pub async fn list(&self) -> Result<Vec<CipherView>, GetCipherError> {
-        let key_store = self.client.internal.get_key_store();
-        let repository = self.get_repository()?;
-
-        list_ciphers(key_store, repository.as_ref()).await
-    }
-
-    /// Get all ciphers from state and decrypt them, returning both successes and failures.
-    /// This method will not fail when some ciphers fail to decrypt, allowing for graceful
-    /// handling of corrupted or problematic cipher data.
-    pub async fn list_with_failures(&self) -> Result<DecryptCipherListResult, GetCipherError> {
-        let key_store = self.client.internal.get_key_store();
-        let repository = self.get_repository()?;
-
-        list_ciphers_with_failures(key_store, repository.as_ref()).await
-    }
-
-    /// Get [Cipher] by ID from state and decrypt it to a [CipherView].
-    pub async fn get(&self, cipher_id: &str) -> Result<CipherView, GetCipherError> {
-        let key_store = self.client.internal.get_key_store();
-        let repository = self.get_repository()?;
-
-        get_cipher(key_store, repository.as_ref(), cipher_id).await
-    }
-
-    /// Create a new [Cipher] and save it to the server.
-    pub async fn create(
-        &self,
-        mut request: CipherAddEditRequest,
-    ) -> Result<CipherView, CreateCipherError> {
-        let key_store = self.client.internal.get_key_store();
-        let config = self.client.internal.get_api_configurations().await;
-        let repository = self.get_repository()?;
-
-        request.encrypted_for = self.client.internal.get_user_id();
-
-        create_cipher(key_store, &config.api, repository.as_ref(), request).await
-    }
-
-    /// Edit an existing [Cipher] and save it to the server.
-    pub async fn edit(
-        &self,
-        cipher_id: &str,
-        mut request: CipherAddEditRequest,
-    ) -> Result<CipherView, EditCipherError> {
-        let key_store = self.client.internal.get_key_store();
-        let config = self.client.internal.get_api_configurations().await;
-        let repository = self.get_repository()?;
-
-        request.encrypted_for = self.client.internal.get_user_id();
-
-        edit_cipher(
-            key_store,
-            &config.api,
-            repository.as_ref(),
-            cipher_id,
-            request,
-        )
-        .await
     }
 }
