+++ crates/bitwarden-vault/src/cipher/mod.rs
@@ -5,10 +5,7 @@ pub(crate) mod card;
 pub(crate) mod cipher;
 pub(crate) mod cipher_client;
 pub(crate) mod cipher_permissions;
-pub(crate) mod create;
-pub(crate) mod edit;
 pub(crate) mod field;
-pub(crate) mod get_list;
 pub(crate) mod identity;
 pub(crate) mod linked_id;
 pub(crate) mod local_data;
